EquivariantModels

Documentation for EquivariantModels.

EquivariantModels.CategoricalBasis

EquivariantModels.LinearSearch

EquivariantModels._invmap

EquivariantModels._nlms2b

EquivariantModels.closure

EquivariantModels.degord2spec

EquivariantModels.dropnames

EquivariantModels.get_natural_spec

EquivariantModels.getspec1idx

EquivariantModels.getspecnlm

EquivariantModels.make_nlms_spec

EquivariantModels.rpe_basis

EquivariantModels.specnlm2spec1p

EquivariantModels.CategoricalBasis — Type

CategoricalBasis : defines the discrete 1p basis

\[ \phi_q(u) = \delta{u, U_q},\]

where $U_q, q = 1, \dots, Q$ are finitely many possible values that the variable $u$ may take. Suppose, e.g., we allow the values [:a, :b, :c], then

P = CategoricalBasis([:a, :b, :c]; varsym = :u, idxsym = :q)
evaluate(P, State(u = :a))     # Bool[1, 0, 0]
evaluate(P, State(u = :b))     # Bool[0, 1, 0]
evaluate(P, State(u = :c))     # Bool[0, 0, 1]

If we evaluate it with an unknown state we get an error:

evaluate(P, State(u = :x))   
# Error : val = x not found in this list

Warning : the list of categories is internally stored as an SVector which means that lookup scales linearly with the number of categories

EquivariantModels.LinearSearch — Method

LinearSearch(arr, t) Return the index of t in arr, note that we use sort(t) == sort(x) instead of t == x

EquivariantModels._invmap — Method

_invmap(a::AbstractVector) Return a dictionary that maps the elements of a to their indices

EquivariantModels._nlms2b — Method

_nlms2b(nn, ll, mm) Return a list of NamedTuples of (n, l, m) from nn, ll, mm

EquivariantModels.closure — Method

closure(specnlm,filter) Make a specnlm to be a "complete" set to be symmetrised w.r.t to the filter

EquivariantModels.degord2spec — Method

degord2spec(;totaldegree, order, Lmax, radialbasis = legendrebasis, wL = 1, islong = true) Return a list of AA specifications and A specifications

EquivariantModels.dropnames — Method

dropnames(namedtuple::NamedTuple, names::Tuple{Vararg{Symbol}}) Return a new NamedTuple with tuples elements in "names" dropped

EquivariantModels.get_natural_spec — Method

getnaturalspec(spec, spec1p) Get a readible version of spec

EquivariantModels.getspec1idx — Method

getspec1idx(spec1, spec_Rnl, bYlm) Return a vector of tuples of indices of spec1 w.r.t actual indices (i.e. 1, 2, 3, ...) of bRnl and bYlm

EquivariantModels.getspecnlm — Method

getspecnlm(spec1p, spec) Get a readible version of spec

EquivariantModels.make_nlms_spec — Method

makenlmsspec(bRnl, bYlm) Return a vector of tuples of indices of spec1 w.r.t naural indices (i.e. (n = ..., l = ..., m = ...) ) of bRnl and bYlm

EquivariantModels.rpe_basis — Method

rpiA2Bmatrix(cgen::Union{Rot3DCoeffs{L,T},Rot3DCoeffsreal{L,T},Rot3DCoeffs_long{L,T}}, spec::Vector{Vector{NamedTuple}}) Return a sparse matrix for symmetrisation of AA basis of spec with equivariance specified by cgen

EquivariantModels.specnlm2spec1p — Method

specnlm2spec1p(spec_nlm) From a list of AA specifications to all A specifications needed